+++ src/converters/piral-ng/src/startup.ts
@@ -73,10 +73,10 @@ if (process.env.NODE_ENV === 'development') {
       console.log('Running in outdated mode (Angular 5-8)');
     },
     current() {
-      console.log('Running in current mode (Angular 9-11)');
+      console.log('Running in current mode (Angular 9-13)');
     },
     next() {
-      console.log('Running in next mode (Angular 12)');
+      console.log('Running in next mode (Angular 14)');
     },
     unknown() {
       console.log('Running with an unknown version of Angular');
@@ -92,7 +92,9 @@ if (process.env.NODE_ENV === 'development') {
     v9: versionHandlers.current,
     v10: versionHandlers.current,
     v11: versionHandlers.current,
-    v12: versionHandlers.next,
+    v12: versionHandlers.current,
+    v13: versionHandlers.current,
+    v14: versionHandlers.next,
   };
 
   const handler = getVersionHandler(versions) || versionHandlers.unknown;