+++ src/tooling/piral-cli/src/common/index.ts
@@ -19,6 +19,7 @@ export * from './merge';
 export * from './npm';
 export * from './pack';
 export * from './package';
+export * from './parallel';
 export * from './patcher';
 export * from './patches';
 export * from './port';